@@ -13,12 +13,8 @@ from .robust_exchange import RobustExchange
 from .robust_queue import RobustQueue
 
 
-try:
-    from importlib.metadata import Distribution
-    __version__ = Distribution.from_name("aio-pika").version
-except ImportError:
-    import pkg_resources
-    __version__ = pkg_resources.get_distribution("aio-pika").version
+from importlib.metadata import Distribution
+__version__ = Distribution.from_name("aio-pika").version
 
 
 __all__ = (
