@@ -2,9 +2,8 @@ import asyncio
 import warnings
 from collections import defaultdict
 from itertools import chain
-from typing import Any, DefaultDict, Dict, MutableSet, Optional, Type, Union
+from typing import Any, DefaultDict, Dict, Optional, Set, Type, Union
 from warnings import warn
-from weakref import WeakSet
 
 import aiormq
 
@@ -32,8 +31,8 @@ class RobustChannel(Channel, AbstractRobustChannel):    # type: ignore
 
     RESTORE_RETRY_DELAY: int = 2
 
-    _exchanges: DefaultDict[str, MutableSet[AbstractRobustExchange]]
-    _queues: DefaultDict[str, MutableSet[RobustQueue]]
+    _exchanges: DefaultDict[str, Set[AbstractRobustExchange]]
+    _queues: DefaultDict[str, Set[RobustQueue]]
     default_exchange: RobustExchange
 
     def __init__(
@@ -61,15 +60,16 @@ class RobustChannel(Channel, AbstractRobustChannel):    # type: ignore
             on_return_raises=on_return_raises,
         )
 
-        self._exchanges = defaultdict(WeakSet)
-        self._queues = defaultdict(WeakSet)
+        self._exchanges = defaultdict(set)
+        self._queues = defaultdict(set)
         self._prefetch_count: int = 0
         self._prefetch_size: int = 0
         self._global_qos: bool = False
-        self.reopen_callbacks: CallbackCollection = CallbackCollection(self)
+        self.reopen_callbacks = CallbackCollection(self)
         self.__restore_lock = asyncio.Lock()
         self.__restored = asyncio.Event()
-        self.close_callbacks.add(self.__close_callback)
+
+        self.close_callbacks.remove(self._set_closed_callback)
 
     async def ready(self) -> None:
         await self._connection.ready()
@@ -94,25 +94,42 @@ class RobustChannel(Channel, AbstractRobustChannel):    # type: ignore
             await self.reopen()
             self.__restored.set()
 
-    async def __close_callback(self, _: Any, exc: BaseException) -> None:
+    async def _on_close(
+        self,
+        closing: asyncio.Future
+    ) -> Optional[BaseException]:
+        exc = await super()._on_close(closing)
+
         if isinstance(exc, asyncio.CancelledError):
             # This happens only if the channel is forced to close from the
             # outside, for example, if the connection is closed.
             # Of course, here you need to exit from this function
             # as soon as possible and to avoid a recovery attempt.
             self.__restored.clear()
-            return
+            if not self._closed.done():
+                self._closed.set_result(True)
+            return exc
 
         in_restore_state = not self.__restored.is_set()
         self.__restored.clear()
 
-        if self._closed or in_restore_state:
-            return
+        if self._closed.done() or in_restore_state:
+            return exc
 
         await self.restore()
 
-    async def _open(self) -> None:
-        await super()._open()
+        return exc
+
+    async def close(
+        self,
+        exc: Optional[aiormq.abc.ExceptionType] = None,
+    ) -> None:
+        # Avoid recovery when channel is explicitely closed using this method
+        self.__restored.clear()
+        await super().close(exc)
+
+    async def reopen(self) -> None:
+        await super().reopen()
         await self.reopen_callbacks()
 
     async def _on_open(self) -> None:
@@ -177,6 +194,11 @@ class RobustChannel(Channel, AbstractRobustChannel):    # type: ignore
         timeout: TimeoutType = None,
         robust: bool = True,
     ) -> AbstractRobustExchange:
+        """
+        :param robust: If True, the exchange will be re-declared during
+        reconnection.
+        Set to False for temporary exchanges that should not be restored.
+        """
         await self.ready()
         exchange = (
             await super().declare_exchange(
@@ -226,6 +248,11 @@ class RobustChannel(Channel, AbstractRobustChannel):    # type: ignore
         timeout: TimeoutType = None,
         robust: bool = True,
     ) -> AbstractRobustQueue:
+        """
+        :param robust: If True, the queue will be re-declared during
+        reconnection.
+        Set to False for temporary queues that should not be restored.
+        """
         await self.ready()
         queue: RobustQueue = await super().declare_queue(   # type: ignore
             name=name,
