@@ -3,7 +3,7 @@ import itertools
 import logging
 from contextlib import suppress
 from functools import partial
-from typing import Callable, List, Type
+from typing import Callable, List, Type, Optional
 
 import aiomisc
 import aiormq.exceptions
@@ -13,8 +13,8 @@ from aiomisc_pytest import TCPProxy  # type: ignore
 from yarl import URL
 
 import aio_pika
-from aio_pika.abc import AbstractRobustChannel
-from aio_pika.exceptions import QueueEmpty
+from aio_pika.abc import AbstractRobustChannel, AbstractRobustConnection
+from aio_pika.exceptions import QueueEmpty, CONNECTION_EXCEPTIONS
 from aio_pika.message import Message
 from aio_pika.robust_channel import RobustChannel
 from aio_pika.robust_connection import RobustConnection
@@ -109,7 +109,7 @@ async def test_revive_passive_queue_on_reconnect(
     reconnect_event = asyncio.Event()
     reconnect_count = 0
 
-    def reconnect_callback(conn):
+    def reconnect_callback(conn: Optional[AbstractRobustConnection]):
         nonlocal reconnect_count
         reconnect_count += 1
         reconnect_event.set()
@@ -565,6 +565,7 @@ async def test_channel_reconnect_after_5kb(
 
         assert messages
 
+    assert on_reconnect.is_set()
     await connection.close()
     await direct_connection.close()
 
@@ -666,7 +667,7 @@ async def test_channel_reconnect_stairway(
             try:
                 await channel.set_qos(prefetch_count=1)
                 break
-            except aiormq.ChannelInvalidStateError:
+            except CONNECTION_EXCEPTIONS:
                 await asyncio.sleep(0.1)
                 continue
 
@@ -689,5 +690,6 @@ async def test_channel_reconnect_stairway(
 
         assert messages
 
+    assert on_reconnect.is_set()
     await connection.close()
     await direct_connection.close()
